refactor server config and db connection
