+++ backend/utils/middleware.js
@@ -25,13 +25,6 @@ const errorHandler = (error, request, response, next) => {
   case 'TokenExpiredError':
     return response.status(401).json({ error: 'token expired' })
   }
-  // if (error.name === 'CastError') {
-  //   return response.status(400).send({ error: 'malformatted id' })
-  // } else if (error.name === 'ValidationError') {
-  //   return response.status(400).json({ error: error.message })
-  // } else if (error.name === 'JsonWebTokenError') {
-  //   return response.status(401).json({ error: 'invalid token' })
-  // } else if (errorname)
 
   next(error)
 }
